+++ src/bctklib/persistence/ICheckpointStore.cs
@@ -1,0 +1,10 @@
+using System;
+using Neo.Persistence;
+
+namespace Neo.BlockchainToolkit.Persistence
+{
+    public interface ICheckpointStore : IReadOnlyStore
+    {
+        ProtocolSettings Settings { get; }
+    }
+}