+++ src/bctklib/persistence/NullCheckpointStore.cs
@@ -1,0 +1,36 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Neo.BlockchainToolkit.Models;
+using Neo.Persistence;
+
+namespace Neo.BlockchainToolkit.Persistence
+{
+    public class NullCheckpointStore : ICheckpointStore
+    {
+        public ProtocolSettings Settings { get; }
+
+        public NullCheckpointStore(ExpressChain? chain)
+            : this(chain?.Network, chain?.AddressVersion)
+        {
+        }
+
+        public NullCheckpointStore(uint? network = null, byte? addressVersion = null)
+        {
+            this.Settings = ProtocolSettings.Default with
+            {
+                Network = network ?? ProtocolSettings.Default.Network,
+                AddressVersion = addressVersion ?? ProtocolSettings.Default.AddressVersion,
+            };
+        }
+
+        public IEnumerable<(byte[] Key, byte[]? Value)> Seek(byte[] key, SeekDirection direction)
+            // See note in Extensions.Seek(RocksDb, ColumnFamilyHandle, ReadOnlySpan<byte>, SeekDirection, ReadOptions?)
+            // regarding this InvalidOperationException 
+            => (key.Length == 0 && direction == SeekDirection.Backward)
+                ? throw new InvalidOperationException("https://github.com/neo-project/neo/issues/2634")
+                : Enumerable.Empty<(byte[], byte[]?)>();
+        public byte[]? TryGet(byte[] key) => null;
+        public bool Contains(byte[] key) => false;
+    }
+}