+++ test/test.bctklib/ContractParameterParserTest.cs
@@ -10,7 +10,7 @@ using Neo.SmartContract;
 using Xunit;
 
 
-namespace test.bctklib3
+namespace test.bctklib
 {
     public class ContractParameterParserTest
     {