+++ test/test.bctklib/RpcClientExtensionsTest.cs
@@ -1,11 +1,12 @@
 using System;
 using Neo;
+using Neo.BlockchainToolkit;
 using Neo.BlockchainToolkit.Persistence;
 using Neo.IO.Json;
 using Neo.Network.RPC;
 using Xunit;
 
-namespace test.bctklib3
+namespace test.bctklib
 {
     public class RpcClientExtensionsTest
     {